Projects/dealership-verification/requirements.txt
flask==3.0.2
pandas==2.2.1
werkzeug==3.0.1
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
requests==2.31.0
lxml==5.1.0
html5lib==1.1
filelock==3.13.1
pytest==7.4.4
tqdm==4.66.2
tenacity==8.2.3
rich==13.7.0
pythonanywhere==0.12.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
pymysql==1.1.0
flask-sqlalchemy==3.1.1
redis==5.0.1
marshmallow==3.20.1

# New dependencies for hybrid extractor
undetected-chromedriver>=3.5.0
firecrawl-py>=1.13.0 